fixes breaking createWiki.ts into functions and hooks
